ai-service/requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
openai
python-dotenv
